my-pets - move post-dog-profile into _lib/actions
